% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logint.R
\name{logint}
\alias{logint}
\title{Logistic Regression Interpretation for One Two-Way Interaction}
\usage{
logint(formula, continuous_var, categorical_var, data)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{continuous_var}{name of the continuous variable in the interaction of interest}

\item{categorical_var}{name of the categorical variable in the interaction of interest}

\item{data}{The name of the dataset}
}
\value{
string object delineating interpretation of interaction effects
}
\description{
a function that returns the interpretation for a logistic regression model with one two-way interaction.
}
\examples{
logint(stroke ~ work_type*age, continuous_var = "age", categorical_var = "work_type", data=test_data)
}
