% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logint3.R
\name{logint3}
\alias{logint3}
\title{Logistic Regression Interpretation for One Two-Way Interaction}
\usage{
logint3(
  formula,
  variable1,
  variable2,
  variable1_type,
  variable2_type,
  data,
  sigfig = 4
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{variable1}{name of the first variable in the interaction of interest}

\item{variable2}{name of the second variable in the interaction of interest}

\item{variable1_type}{type of the first variable in the interaction of interest (either "continuous" or "categorical")}

\item{variable2_type}{type of the second variable in the interaction of interest (either "continuous" or "categorical")}

\item{data}{The name of the dataset}

\item{sigfig}{number of significant figures to round to}
}
\value{
`logint3` prints the interpretation of the result directly to the console using "cat()"
}
\description{
a function that returns the interpretation for a logistic regression model with one two-way interaction.
}
\examples{
## Example 1: Interaction between a continuous and a categorical variable
formula <- stroke ~ work_type*age
variable1 <- "age"
variable2 <- "work_type"
variable1_type <- "continuous"
variable2_type <- "categorical"
data <- test_data
sigfig <- 4
logint3(formula, variable1, variable2, variable1_type, variable2_type, data, sigfig)
## Example 2: Interaction between two continuous variables
formula <- stroke ~ age*avg_glucose_level
variable1 <- "age"
variable2 <- "avg_glucose_level"
variable1_type <- "continuous"
variable2_type <- "continuous"
data <- test_data
sigfig <- 4
logint3(formula, variable1, variable2, variable1_type, variable2_type, data, sigfig)
## Example 3: Interaction between two categorical variables
formula <- stroke ~ work_type*Residence_type
variable1 <- "work_type"
variable2 <- "Residence_type"
variable1_type <- "categorical"
variable2_type <- "categorical"
data <- test_data
sigfig <- 4
logint3(formula, variable1, variable2, variable1_type, variable2_type, data, sigfig)
}
