% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logint_catbycat.R
\name{logint_catbycat}
\alias{logint_catbycat}
\title{Logistic Regression Interpretation for Categorical by Categorical Interaction}
\usage{
logint_catbycat(formula, variable1, variable2, data, sigfig = 4)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{variable1}{The name of the first categorical variable in the interaction}

\item{variable2}{The name of the second categorical variable in the interaction}

\item{data}{The name of the dataset}

\item{sigfig}{number of significant figures to round to}
}
\value{
`logint_catbycat` prints the interpretation of the result directly to the console using "cat()"
}
\description{
a function that returns the interpretation for a logistic regression model with one two-way interaction of a categorical and a categorical variable.
}
\examples{
formula <- stroke ~ work_type*Residence_type
variable1 <- "Residence_type"
variable2 <- "work_type"
data <- test_data
sigfig <- 4
logint_contbycat(formula, continuous_var, categorical_var, data, sigfig)
}
