% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis.R
\name{logis}
\alias{logis}
\title{Logistic Regression Interpretation Without Interaction}
\usage{
logis(formula, variable_interest, variable_type, data, sigfig = 4)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{variable_interest}{a character string specifying the variable of interest in the model.
It must match one of the predictor variables used in the formula.}

\item{variable_type}{a character string specifying the type of the variable of interest.
Acceptable values are "continuous" for numeric variables or "categorical" for factor variables.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which glm is called.}

\item{sigfig}{integer indicating the number of significant digits to be used.}
}
\value{
`logis` prints the interpretation of the result directly to the console using "cat()"
}
\description{
a function that returns the interpretation of logistic regression results
}
\examples{
## Example 1: Continuous variable
formula <- stroke ~ gender + age + hypertension + heart_disease + avg_glucose_level + smoking_status
variable_interest <- "age"
variable_type <- "continuous"
data <- test_data
sigfig <- 4
logis(formula, variable_interest, variable_type, data, sigfig)
## Example 2: Categorical variable
formula <- stroke~gender+age+hypertension+heart_disease+avg_glucose_level+smoking_status
variable_interest <- "smoking_status"
variable_type <- "categorical"
data <- test_data
sigfig <- 4
logis(formula, variable_interest, variable_type, data, sigfig)
}
