% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiSense-package.R
\docType{package}
\name{logiSense-package}
\alias{logiSense}
\alias{logiSense-package}
\title{logiSense: Logistic Regression Interpretation}
\description{
The package includes functions that return the interpretations for various logistic regression models.
}
\keyword{internal}
