% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logint_contbycont.R
\name{logint_contbycont}
\alias{logint_contbycont}
\title{Logistic Regression Interpretation for Continuous by Continuous Interaction}
\usage{
logint_contbycont(formula, variable1, variable2, data, sigfig = 4)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{variable1}{The name of first continuous variable in the interaction}

\item{variable2}{The name of second continuous variable in the interaction}

\item{data}{The name of the dataset}

\item{sigfig}{number of significant figures to round to}
}
\value{
`logint_contbycont` prints the interpretation of the result directly to the console using "cat()"
}
\description{
a function that returns the interpretation for a logistic regression model with one two-way interaction of two continuous variables.
}
\examples{
formula <- stroke ~ age*avg_glucose_level
variable1 <- "age"
variable2 <- "avg_glucose_level"
data <- test_data
sigfig <- 4
logint_contbycont(formula, variable1, variable2, data, sigfig)
}
