% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiSense-package.R
\docType{package}
\name{logiSense-package}
\alias{logiSense}
\alias{logiSense-package}
\title{logiSense: Logistic Regression Interpretation}
\description{
The package includes a function that returns the interpretation of logistic regression results and a function that returns an interpretation for a logistic regression model with one two-way interaction.
}
\keyword{internal}
